+++ sound-mind-studio/tests/test_filter_configuration_panel.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <QObject>
+
+class FilterConfigurationPanelTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void freshPanelHasAFullyTransparentDefaultConfiguration();
+    void changingAStartSpinBoxUpdatesStop0AndEmitsFilterConfigurationChanged();
+    void changingAnEndSpinBoxUpdatesStop1AndEmitsFilterConfigurationChanged();
+    void setFilterConfigurationSyncsAllEightSpinBoxesWithoutEmitting();
+};